Extract DataMatchCalculatedResult as dividual class
