+++ kernel/data-pipeline/core/src/main/java/org/apache/shardingsphere/data/pipeline/core/check/consistency/DataMatchCalculatedResult.java
@@ -1,0 +1,120 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.data.pipeline.core.check.consistency;
+
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.shardingsphere.data.pipeline.api.check.consistency.DataConsistencyCalculatedResult;
+
+import java.math.BigDecimal;
+import java.sql.Array;
+import java.sql.SQLException;
+import java.sql.SQLXML;
+import java.util.Collection;
+import java.util.Iterator;
+import java.util.Objects;
+import java.util.Optional;
+
+/**
+ * Data match calculated result.
+ */
+@RequiredArgsConstructor
+@Getter
+@Slf4j
+public final class DataMatchCalculatedResult implements DataConsistencyCalculatedResult {
+    
+    @NonNull
+    private final Object maxUniqueKeyValue;
+    
+    private final int recordsCount;
+    
+    private final Collection<Collection<Object>> records;
+    
+    @Override
+    public Optional<Object> getMaxUniqueKeyValue() {
+        return Optional.of(maxUniqueKeyValue);
+    }
+    
+    @SneakyThrows(SQLException.class)
+    @Override
+    public boolean equals(final Object o) {
+        if (null == o) {
+            return false;
+        }
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof DataMatchCalculatedResult)) {
+            log.warn("DataMatchCalculatedResult type not match, o.className={}", o.getClass().getName());
+            return false;
+        }
+        final DataMatchCalculatedResult that = (DataMatchCalculatedResult) o;
+        if (recordsCount != that.recordsCount || !Objects.equals(maxUniqueKeyValue, that.maxUniqueKeyValue)) {
+            log.warn("recordCount or maxUniqueKeyValue not match, recordCount1={}, recordCount2={}, maxUniqueKeyValue1={}, maxUniqueKeyValue2={}",
+                    recordsCount, that.recordsCount, maxUniqueKeyValue, that.maxUniqueKeyValue);
+            return false;
+        }
+        EqualsBuilder equalsBuilder = new EqualsBuilder();
+        Iterator<Collection<Object>> thisIterator = records.iterator();
+        Iterator<Collection<Object>> thatIterator = that.records.iterator();
+        while (thisIterator.hasNext() && thatIterator.hasNext()) {
+            equalsBuilder.reset();
+            Collection<Object> thisNext = thisIterator.next();
+            Collection<Object> thatNext = thatIterator.next();
+            if (thisNext.size() != thatNext.size()) {
+                log.warn("record column size not match, size1={}, size2={}, record1={}, record2={}", thisNext.size(), thatNext.size(), thisNext, thatNext);
+                return false;
+            }
+            Iterator<Object> thisNextIterator = thisNext.iterator();
+            Iterator<Object> thatNextIterator = thatNext.iterator();
+            int columnIndex = 0;
+            while (thisNextIterator.hasNext() && thatNextIterator.hasNext()) {
+                ++columnIndex;
+                Object thisResult = thisNextIterator.next();
+                Object thatResult = thatNextIterator.next();
+                boolean matched;
+                if (thisResult instanceof SQLXML && thatResult instanceof SQLXML) {
+                    matched = ((SQLXML) thisResult).getString().equals(((SQLXML) thatResult).getString());
+                } else if (thisResult instanceof BigDecimal && thatResult instanceof BigDecimal) {
+                    matched = DataConsistencyCheckUtils.isBigDecimalEquals((BigDecimal) thisResult, (BigDecimal) thatResult);
+                } else if (thisResult instanceof Array && thatResult instanceof Array) {
+                    matched = Objects.deepEquals(((Array) thisResult).getArray(), ((Array) thatResult).getArray());
+                } else {
+                    matched = equalsBuilder.append(thisResult, thatResult).isEquals();
+                }
+                if (!matched) {
+                    log.warn("record column value not match, columnIndex={}, value1={}, value2={}, value1.class={}, value2.class={}, record1={}, record2={}", columnIndex, thisResult, thatResult,
+                            null != thisResult ? thisResult.getClass().getName() : "", null != thatResult ? thatResult.getClass().getName() : "",
+                            thisNext, thatNext);
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
+    
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder(17, 37).append(getMaxUniqueKeyValue().orElse(null)).append(getRecordsCount()).append(getRecords()).toHashCode();
+    }
+}
